Replace keycodes with curses constants and ord()
Hardcoding the char keycodes seemed to be a very difficult to follow
method of tracking specific character presses. Switched the code to use
curses constants and the built-in ord() function. Due to the easier
code style, the associated comments are also removed since it is no
longer unclear what the specific char keycodes are.